Refactor gulp tasks
