--- .bowerrc
@@ -1,5 +1,0 @@
-{
-  "scripts": {
-    "postinstall": "./.bower-postinstall.sh"
-  }
-}
