--- tasks/clean-js.js
@@ -1,14 +1,0 @@
-/* jshint node: true */
-
-'use strict';
-
-var gulp = require('gulp');
-var paths = require('./_paths');
-var del = require('del');
-
-gulp.task('clean-js', function(cb) {
-  // Ensure the files are deleted before calling next task
-  del(paths.dest + 'javascripts').then(function () {
-    cb();
-  });
-});
