+++ tasks/images.js
@@ -3,14 +3,12 @@
 'use strict';
 
 var gulp = require('gulp');
-var paths = require('./_paths');
+var paths = require('../_paths');
 var imagemin = require('gulp-imagemin');
 
-gulp.task('images', ['clean-images'], function() {
-  var images = paths.images.concat(paths.vendorImages);
-
+gulp.task('images', ['clean:images'], function() {
   gulp
-    .src(images)
+    .src(paths.images)
     .pipe(imagemin({ optimizationLevel: 5 }))
     .pipe(gulp.dest(paths.dest + 'images'));
 });
