+++ tasks/minify-scripts.js
@@ -3,7 +3,7 @@
 'use strict';
 
 var gulp = require('gulp');
-var paths = require('./_paths');
+var paths = require('../_paths');
 var uglify = require('gulp-uglify');
 var rename = require('gulp-rename');
